+++ Gemfile
@@ -3,13 +3,14 @@ source 'https://rubygems.org'
 gemspec :development_group => :test
 
 # gem 'cqm-models', '~> 0.8.4'
-# gem 'mongoid', '~> 5.0.0'
-gem 'mongoid'
+gem 'mongoid', '~> 5.0.0'
 gem 'rubyzip'
 gem 'typhoeus'
-gem 'cqm-models', :path => '../cqm-models'
+
+gem 'cqm-models', git: 'https://github.com/projecttacoma/cqm-models.git', branch: 'measure_object_updates_rebase'
 
 
+# gem 'cqm-models', :path => '../cqm-models'
 
 group :development, :test do
   gem 'bundler-audit'
