+++ lib/hqmf-parser.rb
@@ -7,6 +7,7 @@ require 'ostruct'
 require_relative 'util/counter.rb'
 require_relative 'util/code_system_helper'
 require_relative 'util/hqmf_template_helper'
+require_relative 'util/vsac_api'
 
 require_relative 'hqmf-model/utilities.rb'
 
@@ -115,13 +116,11 @@ require_relative 'qrda-import/patient_importer.rb'
 require_relative 'ext/data_element.rb'
 require_relative 'ext/code.rb'
 
-
-require_relative 'measureloader/cql_loader.rb'
-require_relative 'measureloader/elm_dependency_finder.rb'
-require_relative 'measureloader/elm_parser.rb'
-require_relative 'measureloader/exceptions.rb'
-require_relative 'measureloader/hqmf_measure_loader.rb'
-require_relative 'measureloader/mat_measure_files.rb'
-require_relative 'measureloader/value_set_helpers.rb'
-require_relative 'measureloader/value_set_loader.rb'
-require_relative 'measureloader/vsacapitemp.rb'
+require_relative 'measure-loader/cql_loader'
+require_relative 'measure-loader/elm_dependency_finder'
+require_relative 'measure-loader/elm_parser'
+require_relative 'measure-loader/exceptions'
+require_relative 'measure-loader/hqmf_measure_loader'
+require_relative 'measure-loader/mat_measure_files'
+require_relative 'measure-loader/value_set_helpers'
+require_relative 'measure-loader/value_set_loader'
