+++ lib/measure-loader/cql_loader.rb
@@ -17,12 +17,13 @@ module Measures
         measure = create_measure(measure_files)
       end
 
-      measure.package = CQM::MeasurePackage.new(file: BSON::Binary.new(measure_zip.read()))
+      measure.package = CQM::MeasurePackage.new(file: BSON::Binary.new(measure_zip.read))
       measures = component_measures << measure
       return measures
     end
 
     private
+
     def create_measure_and_components(measure_files)
       add_component_cql_library_files_to_composite_measure_files(measure_files) 
       component_measures = measure_files.components.map { |comp_files| create_measure(comp_files) }
@@ -32,22 +33,20 @@ module Measures
         component_measure.component = true
       end
       measure = create_measure(measure_files)
-      measure.component_hqmf_set_ids = component_measures.map { |cm| cm.hqmf_set_id }
+      measure.component_hqmf_set_ids = component_measures.map(&:hqmf_set_id)
       return measure, component_measures
     end
 
-    private
     def add_component_cql_library_files_to_composite_measure_files(measure_files)
-      component_cql_library_files = measure_files.components.flat_map { |cm| cm.cql_libraries }
+      component_cql_library_files = measure_files.components.flat_map(&:cql_libraries)
       measure_files.cql_libraries.push(*component_cql_library_files)
       measure_files.cql_libraries.uniq! { |cla| cla.id + cla.version }
       return nil
     end
 
     # Creates and returns a measure 
-    private
     def create_measure(measure_files)
-      hqmf_model = HQMF::Parser::V2CQLParser.new().parse(measure_files.hqmf_xml)
+      hqmf_model = HQMF::Parser::V2CQLParser.new.parse(measure_files.hqmf_xml)
 
       measure_files.cql_libraries.each { |cql_lib_files| modify_elm_vs_stuff(cql_lib_files.elm) }
       cql_libraries = create_cql_libraries(measure_files.cql_libraries, hqmf_model.cql_measure_library)
@@ -60,23 +59,21 @@ module Measures
 
       measure = create_measure_from_hqmf_model_hash(hqmf_model_hash)
       measure.cql_libraries = cql_libraries
-      vs_items[:value_set_models].each { |vsm| measure.value_sets.push (vsm) }
+      vs_items[:value_set_models].each { |vsm| measure.value_sets.push vsm }
       measure.composite = measure_files.components.present?
 
       return measure
     end
 
-    private
     def create_measure_from_hqmf_model_hash(hqmf_model_hash)
-      measure_scoring = if @measure_details[:continuous_variable] then 'CONTINUOUS_VARIABLE' else 'PROPORTION' end
+      measure_scoring = @measure_details[:continuous_variable] ? 'CONTINUOUS_VARIABLE' : 'PROPORTION'
       measure = HQMFMeasureLoader.create_measure_model(hqmf_model_hash, measure_scoring)
       measure.measure_scoring = measure_scoring
-      measure.calculation_method = if @measure_details[:episode_of_care] then 'EPISODE_OF_CARE' else 'PATIENT' end
+      measure.calculation_method = @measure_details[:episode_of_care] ? 'EPISODE_OF_CARE' : 'PATIENT'
       measure.calculate_sdes = @measure_details[:calculate_sdes]
       return measure
     end
 
-    private
     def create_cql_libraries(cql_library_files, main_cql_lib)
       cql_statement_dependencies_all_libs = ElmDependencyFinder.find_dependencies(cql_library_files, main_cql_lib)
       
@@ -88,7 +85,6 @@ module Measures
       return cql_libraries
     end
 
-    private
     def modelize_cql_library(cql_lib_files, cql_statement_dependencies, is_main_cql_lib)
       return CQM::CQLLibrary.new(
         library_name: cql_lib_files.id,
@@ -101,7 +97,6 @@ module Measures
       )
     end
 
-    private
     def modelize_cql_statement_dependencies(cql_statment_deps)
       return cql_statment_deps.map do |name, refs|
         refs = refs.map { |ref| CQM::StatementReference.new(ref) }
@@ -109,7 +104,6 @@ module Measures
       end
     end
 
-    private
     def modify_elm_vs_stuff(elm)
       ValueSetHelpers.remove_urnoid(elm)
       ValueSetHelpers.modify_value_set_versions(elm)
