+++ lib/measure-loader/elm_dependency_finder.rb
@@ -1,9 +1,9 @@
 module Measures
-  class ElmDependencyFinder
+  module ElmDependencyFinder
     class << self
 
       def find_dependencies(cql_library_files, main_cql_library_id)
-        elms = cql_library_files.map{ |lib_files| lib_files.elm }
+        elms = cql_library_files.map(&:elm)
         all_elms_dep_map = Hash[elms.map { |elm| [elm_id(elm), make_statement_deps_for_elm(elm)] }]
         needed_deps_map = Hash[elms.map { |elm| [elm_id(elm), {}] }]
 
@@ -20,8 +20,6 @@ module Measures
         return elm['library']['identifier']['id']
       end
 
-      private
-
       def make_library_alias_to_path_hash(elm)
         lib_alias_to_path = { nil => elm_id(elm) } # nil value used for statements without libraryName
         (elm.dig('library','includes','def') || []).each do |library_hash|
@@ -30,42 +28,36 @@ module Measures
         return lib_alias_to_path
       end
 
-      private
-
       def make_statement_deps_for_elm(elm)
         deps = {}
         lib_alias_to_path = make_library_alias_to_path_hash(elm)
         make_statement_deps_for_elm_helper(elm, nil, deps, lib_alias_to_path)
-        deps.each_value {|arr| arr.uniq!}
+        deps.each_value(&:uniq!)
         return deps
       end
 
-      private
-
       def make_statement_deps_for_elm_helper(obj, parent_name, deps, lib_alias_to_path)
-        if obj.kind_of? Array
+        if obj.is_a? Array
           obj.each { |el| make_statement_deps_for_elm_helper(el, parent_name, deps, lib_alias_to_path) }
-        elsif obj.kind_of? Hash
+        elsif obj.is_a? Hash
           if obj['type'].in?(['ExpressionRef', 'FunctionRef']) && parent_name != 'Patient'
-            dep = { :library_name => lib_alias_to_path[obj['libraryName']], :statement_name => obj['name'] }
+            dep = { library_name: lib_alias_to_path[obj['libraryName']], statement_name: obj['name'] }
             deps[parent_name] << dep
-          elsif obj.has_key?('name') && obj.has_key?('expression')
+          elsif obj.key?('name') && obj.key?('expression')
             parent_name = obj['name']
-            deps[parent_name] = [] unless deps.has_key?('parent_name')
+            deps[parent_name] = [] unless deps.key?('parent_name')
           end
           obj.each_pair do |k,v|
             make_statement_deps_for_elm_helper(v, parent_name, deps, lib_alias_to_path) unless k == 'annotation'
           end
         end
       end
-
-      private
       
       def deep_add_external_library_deps(statement, needed_deps_map, all_elms_dep_map)
         s_library = statement[:library_name]
         s_name = statement[:statement_name]
 
-        return if !needed_deps_map.dig(s_library, s_name).nil? # return if key already exists
+        return unless needed_deps_map.dig(s_library, s_name).nil? # return if key already exists
 
         deps_to_add = all_elms_dep_map[s_library][s_name]
         needed_deps_map.deep_merge!(s_library => { s_name => deps_to_add })
