+++ lib/measure-loader/exceptions.rb
