+++ test/simplecov_init.rb
@@ -1,8 +1,9 @@
 require 'simplecov'
 require 'codecov'
+
 SimpleCov.start do
   add_filter "test/"
   track_files 'lib/**/*.rb'
 end
 
-SimpleCov.formatter = SimpleCov::Formatter::Codecov
+SimpleCov.formatters = SimpleCov::Formatter::MultiFormatter.new([SimpleCov::Formatter::HTMLFormatter,SimpleCov::Formatter::Codecov])
